dashboard/app: fix crash selection for reporting
See the issue for the problem description.
Include repro level into reporting priority,
so that we can order by just it during selection
and ignore ReproC/ReproSyz.

Fixes #634

+++ dashboard/app/entities.go
@@ -111,7 +111,7 @@ type Crash struct {
 	// Custom crash priority for reporting (greater values are higher priority).
 	// For example, a crash in mainline kernel has higher priority than a crash in a side branch.
 	// For historical reasons this is called ReportLen.
-	ReportLen int
+	ReportLen int64
 }
 
 // ReportingState holds dynamic info associated with reporting.
